+++ tests/test_json.py
@@ -148,9 +148,6 @@ def test_config() -> None:
             "ImageVersion": "5",
             "Include": [],
             "Incremental": false,
-            "InitrdInclude": [
-                "/foo/bar"
-            ],
             "InitrdPackages": [
                 "clevis"
             ],
@@ -403,7 +400,6 @@ def test_config() -> None:
         image_version="5",
         include=[],
         incremental=False,
-        initrd_include=[Path("/foo/bar"),],
         initrd_packages=["clevis"],
         initrd_volatile_packages=["abc"],
         initrds=[Path("/efi/initrd1"), Path("/efi/initrd2")],